Basic harvest to controller
Just a basic script to leave running while I learn the game. Creeps
have a task stored in memory. At the moment there are two: harvest and
haul. If a creep doesn't have energy it goes to harvest some, when it's
full it hauls it to the room controller and transfers it. Three creeps
looks to be enough for this, as it seems only one can harvest from the
energy source at a time, perhaps because it's in a corner of the walls.