Check if more creeps are needed
When I came back I found a room void of creeps. Creeps have a life
cycle of 1,500 ticks. Turns out they "aged" and died. Their ticksToLive
property reached 0. Note fatigue is related to movement, not time to
live.
In task.haul, assign target dependent on the need for more creeps. In
main, only spawn a creep if more are needed (this code may not be
needed because the spawn wouldn't have the energy anyway, but logically
it belongs).