Minor cleanup
In task.harvest, decide which source should be targeted when assigning
the variable, not every time it is used. There is only one in the room,
but I think this is more clear.
I was going to use a task variable in main, but then assignment
wouldn't persist (task is not a pointer to creep.memory.task).