Absolute positioning test
I didn't like calling find every tick, especially when there is only
one source and one spawn in static positions. I wanted to create a
variable storing the position of these room objects. To quote the
error, "Error: It seems you're trying to use a serialized game object
stored in Memory which is not allowed. Please use `Game.getObjectById`
to retrieve a live object reference instead."
In the end it's just not a good idea. Not only would it mean this code
won't work in a different room, but it just makes sense to find the
source and shouldn't be that computationally expensive.
A log of my testing can be found at
https://gist.github.com/sethten/82b32688a27f20233a2935d025f8c1dd