Fix #6 creeps crowding
Ideally one creep will be mining, one will be traveling between the
room controller and source, and one will be calling
Room.upgradeController(). HARVEST_POWER == 2, but
UPGRADE_CONTROLLER_POWER == 1. A smaller portion of ticks may be wasted
(not have a creep harvesting the source), but this should completely
eliminate the crowding issue.