Add build task
Upgrading the room controller enables us to build some extensions to
use more energy to spawn bigger and more powerful creeps. To start
taking advantage of that, add a 'build' task.
When a creep is full of energy and done harvesting, first check if
there is a creep hauling. If not, go do that. This should be the
priority so the downgrade timer will not fall far. If there is at least
one creep taking care of that, then go put that energy towards building
something!