Calculator for optimal body part set
Add a method to find the best set of body parts for a creep for a
simple harvest task. This is not utilized at the moment because haul
does not bring energy to the extensions and main will not differentiate
creeps with a better body part set for mining in the adjoining room.
Remember to use var calc = require('calc'); when using this!