Better handle different sources
Remove task.harvestAdjoiningSource.js, which should really be (and now
is) in task.harvest.js.
Remove code to help a creep get back to the main room from main.js.
This section should only handle the logic of changing their task, the
task.*.js files handle the execution of that task.
Remove code to figure out which source a creep should target based on
what other creeps are doing.
! Spawn creeps with different body part sets based on their target, and
retain that target in memory. Now creeps can be optimized for their
task.